refactor: rename some components
