+++ packages/sheets-data-validation-ui/src/views/components/formula-input/CheckboxFormulaInput.tsx
