+++ packages/sheets-data-validation-ui/src/views/components/formula-input/index.ts
@@ -15,11 +15,16 @@
  */
 
 import type { FormulaInputType } from '@univerjs/data-validation';
-import { BASE_FORMULA_INPUT_NAME, CHECKBOX_FORMULA_INPUT_NAME, CUSTOM_FORMULA_INPUT_NAME, LIST_FORMULA_INPUT_NAME } from '@univerjs/sheets-data-validation';
-import { BaseFormulaInput } from './base-formula-input';
-import { CheckboxFormulaInput } from './checkbox-formula-input';
-import { CustomFormulaInput } from './custom-formula-input';
-import { ListFormulaInput } from './list-formula-input';
+import {
+    BASE_FORMULA_INPUT_NAME,
+    CHECKBOX_FORMULA_INPUT_NAME,
+    CUSTOM_FORMULA_INPUT_NAME,
+    LIST_FORMULA_INPUT_NAME,
+} from '@univerjs/sheets-data-validation';
+import { BaseFormulaInput } from './BaseFormulaInput';
+import { CheckboxFormulaInput } from './CheckboxFormulaInput';
+import { CustomFormulaInput } from './CustomFormulaInput';
+import { ListFormulaInput } from './ListFormulaInput';
 
 export {
     BASE_FORMULA_INPUT_NAME,
