+++ packages/ui/src/views/components/popup/RectPopup.tsx
@@ -19,11 +19,11 @@ import type { RefObject } from 'react';
 import type { Observable } from 'rxjs';
 import type { IUniverUIConfig } from '../../../controllers/config.schema';
 import { IConfigService } from '@univerjs/core';
-import { useDependency } from '@wendellhu/redi/react-bindings';
 import React, { createContext, useContext, useEffect, useRef } from 'react';
 import { createPortal } from 'react-dom';
 import { useEvent } from '../../../components/hooks/event';
 import { UI_PLUGIN_CONFIG_KEY } from '../../../controllers/config.schema';
+import { useDependency } from '../../../utils/di';
 
 interface IAbsolutePosition {
     left: number;
